Light refactor following peer-review
